+++ CHANGES.md
@@ -14,6 +14,7 @@ Thread safe, reentrant, modifiable.
 - got basic multi-interface charts working
 - added debug output for fcgiwrap.service
 - added legend to multi-interface charts
+- added packet size distribution to 'switch' primitive
 
 # Release 1.3.16 (5-May-26): Physical Cyber Security IV
 - Added installation of MRTG banner images